refactor(frontend): rename Ethereum send form component (#4120)
# Motivation

For consistency, we rename the component `SendForm` to `EthSendForm`.
